apply pr feedback
consolidate ensdb modules; replace pino logger with native console logger; simplify ensdb worker logic
