+++ apps/ensindexer/package.json
@@ -29,15 +29,14 @@
     "@ensnode/ensnode-sdk": "workspace:*",
     "@ensnode/ensrainbow-sdk": "workspace:*",
     "@ensnode/ponder-metadata": "workspace:*",
+    "@ponder/client": "catalog:",
     "caip": "catalog:",
     "date-fns": "catalog:",
     "deepmerge-ts": "^7.1.5",
     "dns-packet": "^5.6.1",
     "drizzle-orm": "catalog:",
     "p-retry": "catalog:",
-    "pino": "catalog:",
     "pg-connection-string": "catalog:",
-    "pg": "8.16.3",
     "hono": "catalog:",
     "ponder": "catalog:",
     "viem": "catalog:",
