+++ apps/ensindexer/ponder/src/ensdb-writer-worker.ts
@@ -4,8 +4,6 @@
  * - Indexing Status
  * into the ENSDb.
  */
-import config from "@/config";
-
 import { secondsToMilliseconds } from "date-fns";
 import pRetry from "p-retry";
 
@@ -18,13 +16,10 @@ import {
 } from "@ensnode/ensnode-sdk";
 
 import { validateENSIndexerPublicConfigCompatibility } from "@/config/compatibility";
-import { EnsDbConnection, EnsDbMutation, EnsDbQuery } from "@/lib/ensdb";
+import { EnsDbClient } from "@/lib/ensdb";
 import { ensIndexerClient, waitForEnsIndexerToBecomeHealthy } from "@/lib/ensindexer";
-import { makeLogger } from "@/lib/logger";
 
 const INDEXING_STATUS_RECORD_UPDATE_INTERVAL: Duration = 1;
-
-const logger = makeLogger("ensdb-writer-worker");
 
 /**
  * ENSDb Writer Worker
@@ -38,35 +33,27 @@ const logger = makeLogger("ensdb-writer-worker");
  *    into ENSDb.
  */
 async function ensDbWriterWorker() {
+  console.log("ENSDb Writer Worker: waiting for ENSIndexer to become healthy.");
+
   // 0. Wait for ENSIndexer to become healthy before running the worker's logic
   await waitForEnsIndexerToBecomeHealthy;
 
+  console.log("ENSDb Writer Worker: ENSIndexer is healthy, starting tasks.");
+
   // 1. Create ENSDb Client
-  const ensDbConnection = new EnsDbConnection();
-  const ensDbClient = ensDbConnection.connect({
-    schemaName: config.databaseSchemaName,
-    poolConfig: {
-      connectionString: config.databaseUrl,
-    },
-  });
-
-  logger.info("ENSDb Client connected");
-
-  // 2. Create ENSDb Query object for read operations
-  const ensDbQuery = new EnsDbQuery(ensDbClient);
-  // 3. Create ENSDb Mutation object for write operations
-  const ensDbMutation = new EnsDbMutation(ensDbClient);
+  const ensDbClient = new EnsDbClient();
 
   /**
    * Handle ENSIndexerPublicConfig Record
    */
   const handleEnsIndexerPublicConfigRecord = async () => {
     // Read stored config and in-memory config.
-    // Note: we wrap read operations in pRetry to ensure all of them are
+    // Note: we wrap each operation in pRetry to ensure all of them can be
     // completed successfully.
-    const [storedConfig, inMemoryConfig] = await pRetry(() =>
-      Promise.all([ensDbQuery.getEnsIndexerPublicConfig(), ensIndexerClient.config()]),
-    );
+    const [storedConfig, inMemoryConfig] = await Promise.all([
+      pRetry(() => ensDbClient.getEnsIndexerPublicConfig()),
+      pRetry(() => ensIndexerClient.config()),
+    ]);
 
     // Validate in-memory config object compatibility with the stored one,
     // if the stored one is available
@@ -74,21 +61,17 @@ async function ensDbWriterWorker() {
       try {
         validateENSIndexerPublicConfigCompatibility(storedConfig, inMemoryConfig);
       } catch (error) {
-        const errorMessage =
-          "In-memory ENSIndexerPublicConfig object is not compatible with its counterpart stored in ENSDb.";
-
-        logger.error(error, errorMessage);
+        const errorMessage = `In-memory ENSIndexerPublicConfig object is not compatible with its counterpart stored in ENSDb.`;
 
         // Throw the error to terminate the ENSIndexer process due to
         // found config incompatibility
-        throw new Error(errorMessage);
+        throw new Error(errorMessage, {
+          cause: error,
+        });
       }
     } else {
       // Upsert ENSIndexerPublicConfig into ENSDb.
-      // Note: we wrap write operation in pRetry to ensure it can complete
-      // successfully, as there will be no other attempt.
-      await pRetry(() => ensDbMutation.upsertEnsIndexerPublicConfig(inMemoryConfig));
-      logger.info("ENSIndexer Public Config successfully stored in ENSDb.");
+      await ensDbClient.upsertEnsIndexerPublicConfig(inMemoryConfig);
     }
   };
 
@@ -110,16 +93,14 @@ async function ensDbWriterWorker() {
 
       // Check if Indexing Status is in expected status.
       if (omnichainSnapshot.omnichainStatus === OmnichainIndexingStatusIds.Unstarted) {
-        throw new Error("Omnichain Status  must be different that 'Unstarted'.");
+        throw new Error("Omnichain Status must be different than 'Unstarted'.");
       }
 
       // Upsert ENSIndexerPublicConfig into ENSDb.
-      await ensDbMutation.upsertIndexingStatus(snapshot);
-
-      logger.info("Indexing Status successfully stored in ENSDb.");
+      await ensDbClient.upsertIndexingStatus(snapshot);
     } catch (error) {
       // Do nothing about this error, but having it logged.
-      logger.error(error, "Could not upsert Indexing Status record");
+      console.error(error, "Could not upsert Indexing Status record");
     } finally {
       // Regardless of current iteration result,
       // schedule the next callback to handle Indexing Status Record.
@@ -131,12 +112,20 @@ async function ensDbWriterWorker() {
   };
 
   // 4. Handle ENSIndexer Public Config just once.
-  await handleEnsIndexerPublicConfigRecord();
+  console.log("Task: store ENSIndexer Public Config in ENSDb.");
+  await handleEnsIndexerPublicConfigRecord().then(() =>
+    console.log("ENSIndexer Public Config successfully stored in ENSDb."),
+  );
 
   // 5. Handle Indexing Status on recurring basis.
-  await handleIndexingStatusRecordRecursively();
+  console.log("Task: store Indexing Status in ENSDb.");
+  await handleIndexingStatusRecordRecursively().then(() =>
+    console.log("Indexing Status successfully stored in ENSDb."),
+  );
 }
 
 // Run ENSDb Writer Worker in a non-blocking way to
 // allow database migrations to proceed in the background.
-setTimeout(ensDbWriterWorker, 0);
+ensDbWriterWorker().catch((error) =>
+  console.error("ENSDb Writer Worker failed to perform its tasks", error),
+);
