+++ apps/ensindexer/src/lib/ensdb/index.ts
@@ -1,3 +1,1 @@
-export * from "./ensdb-connection";
-export * from "./ensdb-mutation";
-export * from "./ensdb-query";
+export * from "./ensdb-client";
