+++ pnpm-workspace.yaml
@@ -8,6 +8,7 @@ catalog:
   "@astrojs/react": ^4.4.1
   "@astrojs/tailwind": ^6.0.2
   "@namehash/namekit-react": 0.12.0
+  "@ponder/client": 0.13.14
   "@ponder/utils": 0.2.14
   "@testing-library/react": ^16.3.0
   "@types/node": 22.18.13
@@ -18,6 +19,7 @@ catalog:
   date-fns: 4.1.0
   drizzle-orm: "=0.41.0"
   hono: ^4.10.2
+  p-memoize: ^8.0.0
   p-retry: ^7.1.1
   pg-connection-string: ^2.9.1
   pino: 10.1.0
